requirements.txt
matplotlib==3.5.3
pandas==1.3.5
Pillow==9.4.0
requests==2.28.2
streamlit==1.19.0
streamlit_lottie==0.0.3
wikipedia==1.4.0
yfinance==0.2.12
